refactor: refactored main screen imports
