+++ lib/features/favorites/presentation/favorite_events_list_screen.dart
