+++ lib/features/main/presentation/main_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:ows_events_mobile/common_widgets/info_icon_button.dart';
 import 'package:ows_events_mobile/common_widgets/logo.dart';
 import 'package:ows_events_mobile/features/events/presentation/events_screen.dart';
-import 'package:ows_events_mobile/features/favourites/presentation/favourites_screen.dart';
+import 'package:ows_events_mobile/features/favorites/presentation/favorite_events_list_screen.dart';
 import 'package:ows_events_mobile/features/main/presentation/main_screen_controller.dart';
 
 class MainScreen extends ConsumerWidget {
@@ -11,7 +11,7 @@ class MainScreen extends ConsumerWidget {
 
   final List<Widget> screens = const [
     EventsScreen(),
-    FavouritesListScreen(),
+    FavoriteEventsListScreen(),
   ];
 
   @override
@@ -36,19 +36,20 @@ class MainScreen extends ConsumerWidget {
               ],
             ),
             body: PageView(
-                controller: controller,
-                children: screens,
-                onPageChanged: (index) =>
-                    ref.read(indexProvider.notifier).value = index,
-                    ),
+              controller: controller,
+              children: screens,
+              onPageChanged: (index) =>
+                  ref.read(indexProvider.notifier).value = index,
+            ),
             bottomNavigationBar: NavigationBar(
               selectedIndex: menuIndex,
               onDestinationSelected: (int index) {
                 ref.read(indexProvider.notifier).value = index;
-                controller.animateToPage(index,
-                    duration: const Duration(seconds: 1),
-                    curve: Curves.easeInOut,
-                    );
+                controller.animateToPage(
+                  index,
+                  duration: const Duration(seconds: 1),
+                  curve: Curves.easeInOut,
+                );
               },
               destinations: const <NavigationDestination>[
                 NavigationDestination(
